mansyntax: make it work on macOS, check reqs locally
- use `gman` if present. This makes it work when the correct `man`
  command is provided via `brew` on macOS.

- move CMake attempts to detect tools necessary to run `mansyntax.sh`
  into the script itself.

- delete CMake TODO to move more test logic into CMake. This would make
  it CMake-specific and require maintaining it separately for each build
  tool. Just use out external script when a POSIX shell is available.

Closes #xxx

+++ tests/mansyntax.sh
@@ -5,6 +5,10 @@ set -e
 #
 # Run syntax checks for all manpages in the documentation tree.
 #
+# Requirement on macOS: brew install man-db
+#
+
+command -v gman >/dev/null 2>&1 && man() { gman "$@"; }
 
 srcdir="${srcdir:-$PWD}"
 dstdir="${builddir:-$PWD}"
@@ -15,7 +19,8 @@ ec=0
 #
 # Only test if suitable man is available
 #
-if man --help | grep -q warnings; then
+if command -v grep >/dev/null 2>&1 && \
+   man --help | grep -q warnings; then
 
   trap 'rm -f "$dstdir/man3"' EXIT
 
@@ -24,14 +29,14 @@ if man --help | grep -q warnings; then
   for manpage in "$mandir"/libssh2_*.*; do
     echo "$manpage"
     warnings=$(LANG=en_US.UTF-8 MANWIDTH=80 man -M "$dstdir" --warnings \
-      -E UTF-8 -l "$manpage" 2>&1 >/dev/null)
+      -E UTF-8 -l "$manpage" >/dev/null 2>&1)
     if [ -n "$warnings" ]; then
       echo "$warnings"
       ec=1
     fi
   done
 else
-  echo "man version not suitable, skipping tests"
+  echo 'Required tool not found, skipping mansyntax tests.'
 fi
 
 exit "$ec"
